+++ lib/vector_number/mathing.rb
@@ -64,12 +64,13 @@ class VectorNumber
     end
 
     # Divide all coefficients by a real number, returning new vector.
-    # This effectively multiplies magnitude by reciprocal of the specified factor.
+    # This effectively multiplies magnitude by reciprocal of +other+.
     # @param other [Integer, Float, Rational, BigDecimal, VectorNumber]
     # @return [VectorNumber]
     # @raise [RangeError] if +other+ is not a number or is not a real number
+    # @raise [ZeroDivisionError] if +other+ is zero
     def /(other)
-      raise RangeError, "can't divide #{self} by #{other}" unless real_number?(other)
+      check_divisibility(other)
 
       other = other.real
       # Prevent integer division, but without loss of accuracy.
@@ -78,16 +79,84 @@ class VectorNumber
       new { _1 / other }
     end
 
+    alias quo /
+
     # Divide all coefficients by a real number using +fdiv+, returning new vector
     # with Float coefficients.
     # @param other [Integer, Float, Rational, BigDecimal, VectorNumber]
     # @return [VectorNumber]
     # @raise [RangeError] if +other+ is not a number or is not a real number
+    # @raise [ZeroDivisionError] if +other+ is zero
     def fdiv(other)
-      raise RangeError, "can't divide #{self} by #{other}" unless real_number?(other)
+      check_divisibility(other)
 
       other = other.real
       new { _1.fdiv(other) }
+    end
+
+    # Divide all coefficients by +other+, rounding results with {#floor}.
+    # This is requal to +(self / other).floor+.
+    # @param other [Integer, Float, Rational, BigDecimal, VectorNumber]
+    # @return [VectorNumber]
+    # @raise [RangeError] if +other+ is not a number or is not a real number
+    # @raise [ZeroDivisionError] if +other+ is zero
+    # @see #divmod
+    # @see #%
+    def div(other)
+      check_divisibility(other)
+
+      new { _1.div(other) }
+    end
+
+    # Return the remainder of dividing self by +other+ as a vector.
+    # This is equal to +self - other * (self/other).floor+.
+    # @param other [Integer, Float, Rational, BigDecimal, VectorNumber]
+    # @return [VectorNumber]
+    # @raise [RangeError] if +other+ is not a number or is not a real number
+    # @raise [ZeroDivisionError] if +other+ is zero
+    # @see #divmod
+    # @see #div
+    # @see #remainder for alternative
+    # @see Numeric#% for examples
+    def %(other)
+      check_divisibility(other)
+
+      new { _1 - other * _1.div(other) }
+    end
+
+    alias modulo %
+
+    # Return the quotient and remainder of dividing self by +other+.
+    # There is no performance benefit compared to calling {#div} and {#%} separately.
+    # @param other [Integer, Float, Rational, BigDecimal, VectorNumber]
+    # @return [Array(VectorNumber, VectorNumber)]
+    # @raise [RangeError] if +other+ is not a number or is not a real number
+    # @raise [ZeroDivisionError] if +other+ is zero
+    # @see #div
+    # @see #%
+    def divmod(other)
+      [div(other), modulo(other)]
+    end
+
+    # Return the alternative remainder of dividing self by +other+ as a vector.
+    # This is equal to +self - other * (self/other).truncate+.
+    # @param other [Integer, Float, Rational, BigDecimal, VectorNumber]
+    # @return [VectorNumber]
+    # @raise [RangeError] if +other+ is not a number or is not a real number
+    # @raise [ZeroDivisionError] if +other+ is zero
+    # @see #% for alternative
+    # @see Numeric#remainder for examples
+    def remainder(other)
+      check_divisibility(other)
+
+      new { _1 - other * (_1 / other).truncate }
+    end
+
+    private
+
+    def check_divisibility(other)
+      raise RangeError, "can't divide #{self} by #{other}", caller unless real_number?(other)
+      raise ZeroDivisionError, "divided by 0", caller if other.zero?
     end
   end
 end